+++ src/components/index.js
@@ -8,10 +8,10 @@ export * from './timeline';
 export * from './task';
 export * from './project';
 export * from './log';
-export * from './text-editor';
 export * from './events';
 export * from './page-container';
 export * from './kanban-board';
 export * from './loader-full-page';
 export * from './loader-generic';
 export * from './organizations';
+export * from './modal';
