--- src/components/text-editor/index.js
@@ -1,1 +1,0 @@
-export * from './TextEditor';
